move heal task assignment, resp processing and revert helpers to sync_v1.go
